+++ src/lib/settings.ts
@@ -30,6 +30,13 @@ export const settingsSchema: SettingSchemaDesc[] = [
     description: "See Engines in OpenAI docs.",
   },
   {
+    key: "chatCompletionEndpoint",
+    type: "string",
+    default: "http://api.openai.com/v1/",
+    title: "OpenAI API Completion Endpoint",
+    description: "The endpoint to use for OpenAI API completion requests. You shouldn't need to change this."
+  },
+  {
     key: "chatPrompt",
     type: "string",
     default: "Do not refer to yourself in your answers. Do not say as an AI language model...",
@@ -42,7 +49,8 @@ export const settingsSchema: SettingSchemaDesc[] = [
     default: 1.0,
     title: "OpenAI Temperature",
     description:
-      "The temperature controls how much randomness is in the output.",
+      "The temperature controls how much randomness is in the output.<br/>"+
+      "You can set a different temperature in your own prompt templates by adding a 'prompt-template' property to the block.",
   },
   {
     key: "openAIMaxTokens",
@@ -99,6 +107,7 @@ export function getOpenaiSettings(): PluginOptions {
     logseq.settings!["dalleImageSize"]
   ) as DalleImageSize;
   const chatPrompt = logseq.settings!["chatPrompt"];
+  const completionEndpoint = logseq.settings!["chatCompletionEndpoint"];
   return {
     apiKey,
     baseUrl,
@@ -108,5 +117,6 @@ export function getOpenaiSettings(): PluginOptions {
     dalleImageSize,
     injectPrefix,
     chatPrompt,
+    completionEndpoint,
   };
 }