+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [1.12.0](https://github.com/briansunter/logseq-plugin-gpt3-openai/compare/v1.11.0...v1.12.0) (2023-06-01)
+
+
+### Features
+
+* **templates:** Allow user to set the temperature for each prompt ([#111](https://github.com/briansunter/logseq-plugin-gpt3-openai/issues/111)) ([8c05660](https://github.com/briansunter/logseq-plugin-gpt3-openai/commit/8c05660a296fe4e4d3b662e7923f9617e0b73425))
+
+# [1.11.0](https://github.com/briansunter/logseq-plugin-gpt3-openai/compare/v1.10.0...v1.11.0) (2023-06-01)
+
+
+### Features
+
+* **settings:** make openai /v1/chat/completion endpoint(basepath) configurable. ([#110](https://github.com/briansunter/logseq-plugin-gpt3-openai/issues/110)) ([b284838](https://github.com/briansunter/logseq-plugin-gpt3-openai/commit/b284838d1a56d590de11bc05f649dfb55770ded6))
+
 # [1.10.0](https://github.com/briansunter/logseq-plugin-gpt3-openai/compare/v1.9.0...v1.10.0) (2023-03-20)